refactor: adjust vote transaction step 1
