+++ src/domains/transaction/components/VoteTransactionType/index.tsx
@@ -1,0 +1,1 @@
+export * from "./VoteTransactionType";
