+++ src/domains/transaction/pages/SendVote/FormStep.tsx
@@ -2,15 +2,14 @@ import { Contracts as ProfilesContracts } from "@ardenthq/sdk-profiles";
 import React, { useMemo } from "react";
 import { useTranslation } from "react-i18next";
 
-import { useFormContext } from "react-hook-form";
 import { SendVoteStepProperties } from "./SendVote.contracts";
 import { FormField, FormLabel } from "@/app/components/Form";
 import { FeeField } from "@/domains/transaction/components/FeeField";
-import { TransactionDetail } from "@/domains/transaction/components/TransactionDetail";
-import { VoteList } from "@/domains/vote/components/VoteList";
 import { StepHeader } from "@/app/components/StepHeader";
-import { SelectAddress } from "@/domains/profile/components/SelectAddress";
-import { SelectNetworkDropdown } from "@/app/components/SelectNetworkDropdown";
+import {DetailTitle, DetailWrapper} from "@/app/components/DetailWrapper";
+import {Address} from "@/app/components/Address";
+import {ThemeIcon} from "@/app/components/Icon";
+import {VoteTransactionType} from "@/domains/transaction/components/VoteTransactionType";
 
 type FormStepProperties = {
 	profile: ProfilesContracts.IProfile;
@@ -18,9 +17,8 @@ type FormStepProperties = {
 	isWalletFieldDisabled?: boolean;
 } & Omit<SendVoteStepProperties, "wallet">;
 
-export const FormStep = ({ unvotes, votes, wallet, profile, network, isWalletFieldDisabled }: FormStepProperties) => {
+export const FormStep = ({ unvotes, votes, wallet, profile, network }: FormStepProperties) => {
 	const { t } = useTranslation();
-	const { setValue } = useFormContext();
 
 	const showFeeInput = useMemo(() => network.chargesZeroFees() === false, [wallet]);
 
@@ -39,66 +37,36 @@ export const FormStep = ({ unvotes, votes, wallet, profile, network, isWalletFie
 	);
 
 	return (
-		<section data-testid="SendVote__form-step">
+		<section data-testid="SendVote__form-step" className="space-y-3 sm:space-y-4">
 			<StepHeader
 				title={t("TRANSACTION.PAGE_VOTE.FORM_STEP.TITLE")}
+				titleIcon={
+					<ThemeIcon dimensions={[24, 24]} lightIcon="SendTransactionLight" darkIcon="SendTransactionDark" />
+				}
 				subtitle={t("TRANSACTION.PAGE_VOTE.FORM_STEP.DESCRIPTION")}
 			/>
 
-			<div className="my-8 space-y-8">
-				<FormField name="network">
-					<FormLabel label={t("COMMON.CRYPTOASSET")} />
-					<SelectNetworkDropdown
-						profile={profile}
-						networks={[network]}
-						selectedNetwork={network}
-						isDisabled
+			<DetailWrapper label={t("TRANSACTION.ADDRESSING")}>
+				<div className="flex w-full items-center justify-between gap-4 space-x-2 sm:justify-start sm:space-x-0">
+					<DetailTitle className="w-auto sm:min-w-28">{t("COMMON.FROM")}</DetailTitle>
+					<Address
+						address={wallet?.address()}
+						walletName={wallet?.alias()}
+						walletNameClass="text-theme-text text-sm leading-[17px] sm:leading-5 sm:text-base"
+						addressClass="text-theme-secondary-500 dark:text-theme-secondary-700 text-sm leading-[17px] sm:leading-5 sm:text-base"
+						wrapperClass="justify-end sm:justify-start"
+						showCopyButton
 					/>
-				</FormField>
+				</div>
+			</DetailWrapper>
 
-				<FormField name="senderAddress">
-					<FormLabel label={t("TRANSACTION.SENDER")} />
-
-					<div data-testid="sender-address">
-						<SelectAddress
-							disabled={isWalletFieldDisabled !== false}
-							wallet={
-								wallet
-									? {
-											address: wallet.address(),
-											network: wallet.network(),
-										}
-									: undefined
-							}
-							wallets={profile.wallets().findByCoinWithNetwork(network.coin(), network.id())}
-							profile={profile}
-							onChange={(address: string) =>
-								setValue("senderAddress", address, { shouldDirty: true, shouldValidate: false })
-							}
-						/>
-					</div>
-				</FormField>
-			</div>
-
-			{unvotes.length > 0 && (
-				<TransactionDetail label={t("TRANSACTION.UNVOTES_COUNT", { count: unvotes.length })}>
-					<VoteList votes={unvotes} currency={wallet?.currency() as string} isNegativeAmount />
-				</TransactionDetail>
-			)}
-
-			{votes.length > 0 && (
-				<TransactionDetail label={t("TRANSACTION.VOTES_COUNT", { count: votes.length })}>
-					<VoteList votes={votes} currency={wallet?.currency() as string} />
-				</TransactionDetail>
-			)}
+			<VoteTransactionType votes={votes} unvotes={unvotes} />
 
 			{showFeeInput && (
-				<TransactionDetail paddingPosition="top">
-					<FormField name="fee" className="flex-1">
-						<FormLabel label={t("TRANSACTION.TRANSACTION_FEE")} />
-						<FeeField type="vote" data={feeTransactionData} network={network} profile={profile} />
-					</FormField>
-				</TransactionDetail>
+				<FormField name="fee" className="flex-1">
+					<FormLabel label={t("TRANSACTION.TRANSACTION_FEE")} />
+					<FeeField type="vote" data={feeTransactionData} network={network} profile={profile} />
+				</FormField>
 			)}
 		</section>
 	);
